==> 05_CDataAnalysis.R <==
##### Prepare data for analysis ---------------------------------------------------------------------------------------------------------------------

# Code author(s): Tom

# Define inverse logit and prepare data for model fitting
if(boot.switch == FALSE){
  
  # Define inverse logit function for later use; need to do this only once
  invlogit <- function(x){exp(x)/(1 + exp(x))}
  
  # Create unique transect as interaction of transect and site
  # First-time only; this bit of code in 06_BootRes does it all other times
  LATR_full <- CData %>% 
    mutate(unique.transect = interaction(transect, site))}


# Set the gamma argument of gam(): gamma > 1 generates smoother fits, less likely to be overfit
gamma <- 1.8





##### Growth model ----------------------------------------------------------------------------------------------------------------------------------

# Prepare a data subset for growth that drops rows missing either t or t1 size data
# Also create log_volume as a new variable because GAM doesn't like functions of variables as variables
LATR_grow <- LATR_full  %>% drop_na(volume_t, volume_t1) %>%
  mutate(log_volume_t = log(volume_t),
         log_volume_t1 = log(volume_t1))

# Create empty list to populate with model results
LATR_gam_models <- list()

# Three candidate models for the mean: size only, size + density, or size, density, and size:density
# Three candidates for variance: size only, size + density, fitted value (all the covariates plus rfx)

# Pilot fits, where sigma depends on initial size only; constant sigma
LATR_gam_models[[1]] <- gam(list(log_volume_t1 ~ s(log_volume_t) + s(unique.transect, bs = "re"), ~ 1),
                            data = LATR_grow, gamma = gamma, family = gaulss())
LATR_gam_models[[2]] <- gam(list(log_volume_t1 ~ s(log_volume_t) + s(weighted.dens) + s(unique.transect, bs = "re"), ~ 1), 
                            data = LATR_grow, gamma = gamma, family = gaulss())                
LATR_gam_models[[3]] <- gam(list(log_volume_t1 ~ s(log_volume_t) + s(weighted.dens) + ti(log_volume_t, weighted.dens) +
                                   s(unique.transect, bs = "re"), ~ 1), 
                            data = LATR_grow, gamma = gamma, family = gaulss()) 
LATR_gam_models[[4]] <- gam(list(log_volume_t1 ~ s(log_volume_t) + s(unique.transect,bs = "re"), ~ s(log_volume_t)), 
                            data = LATR_grow, gamma = gamma, family = gaulss())
LATR_gam_models[[5]] <- gam(list(log_volume_t1 ~ s(log_volume_t) + s(weighted.dens) + s(unique.transect, bs = "re"), ~ s(log_volume_t)), 
                            data = LATR_grow, gamma = gamma, family = gaulss())                
LATR_gam_models[[6]] <- gam(list(log_volume_t1 ~ s(log_volume_t) + s(weighted.dens) + ti(log_volume_t, weighted.dens) +
                                   s(unique.transect, bs = "re"), ~ s(log_volume_t)), 
                            data = LATR_grow, gamma = gamma, family = gaulss()) 

# Fits where sigma depends on both initial size and density
LATR_gam_models[[7]] <- gam(list(log_volume_t1 ~ s(log_volume_t) + s(unique.transect, bs = "re"), ~ s(log_volume_t) + s(weighted.dens)), 
                            data = LATR_grow, gamma = gamma, family = gaulss())
LATR_gam_models[[8]] <- gam(list(log_volume_t1 ~ s(log_volume_t) + s(weighted.dens) + s(unique.transect, bs = "re"),
                                 ~ s(log_volume_t) + s(weighted.dens)), 
                            data = LATR_grow, gamma = gamma, family = gaulss())                
LATR_gam_models[[9]] <- gam(list(log_volume_t1 ~ s(log_volume_t) + s(weighted.dens) + ti(log_volume_t, weighted.dens) +
                                   s(unique.transect, bs = "re"), ~ s(log_volume_t) + s(weighted.dens)), 
                            data = LATR_grow, gamma = gamma, family = gaulss()) 

# Collect model AICs into a single table
grow_aic <- AICtab(LATR_gam_models, base = TRUE, sort = FALSE)

# Set top model as "best"; find intercept for the SD
# Fixing the growth model to this one, so every bootstrap will use this model
# This avoids problems of bad initial conditions in the SGT fits
LATR_grow_best <- LATR_gam_models[[which.min(grow_aic$AIC)]]
LATR_grow_fitted_terms <- predict(LATR_grow_best, type = "terms") 
LATR_grow$pred <- predict.gam(LATR_grow_best, newdata = LATR_grow, exclude = "s(unique.transect)")

# Extract the linear predictor for the mean and SD
LATR_Xp <- predict.gam(LATR_grow_best, type = "lpmatrix")

# Fitted coefficients
LATR_beta <- coef(LATR_grow_best)

# Annoying but necessary index wrangling
# First line where SD coefficients start, second line where lambda, p and q coefficients start
grow_sd_index <- which(as.factor(names(coef(LATR_grow_best))) == "(Intercept).1")
gam_coef_length <- length(coef(LATR_grow_best))

# Now fit SGT using the LP matrix for mean and sigma that GAM fit
# For lambda, I know there is negative skew at small sizes so I am rigging this to pick it up
# sgtLogLik <- function(pars, response){
#   val <- dsgt(x = response, 
#               mu = LATR_Xp[, 1:(grow_sd_index - 1)] %*% pars[1:(grow_sd_index - 1)],
#               sigma = exp(LATR_Xp[, grow_sd_index:gam_coef_length] %*% pars[grow_sd_index:gam_coef_length]),
#               lambda = -invlogit(pars[(gam_coef_length + 1)] + pars[(gam_coef_length + 2)]*LATR_grow$log_volume_t),
#               p = exp(pars[(gam_coef_length + 3)]),
#               q = exp(pars[(gam_coef_length + 4)]),
#               mean.cent = T,
#               var.adj = T,
#               log = T) 
#  return(val)}

# Initial parameter values
# I cut randomness from initial values to make this more reliable for bootstrapping: start = p0*exp(0.2*rnorm(length(p0)))
# p0 <-  c(LATR_beta, -10, 0, 2, 2) 

# Get maximum likelihood
# out <- maxLik(logLik = sgtLogLik, start = p0, response = LATR_grow$log_volume_t1,
#               method = "BHHH", control = list(iterlim = 5000, printLevel = 2), finalHessian = FALSE)
# out <- maxLik(logLik = sgtLogLik, start = out$estimate, response = LATR_grow$log_volume_t1,
#               method = "NM", control = list(iterlim = 5000, printLevel = 1), finalHessian = FALSE)
# out <- maxLik(logLik = sgtLogLik, start = out$estimate, response = LATR_grow$log_volume_t1,
#               method = "BHHH", control = list(iterlim = 5000, printLevel = 2), finalHessian = FALSE)
# out <- maxLik(logLik=sgtLogLik,start=out$estimate,response=LATR_grow$log_volume_t1,
#               method = "BHHH", control = list(iterlim = 5000, printLevel = 2), finalHessian = TRUE) 
#
# Final parameter estimates for the growth model
# coef_grow_best <- out$estimate





##### Flowering probability model -------------------------------------------------------------------------------------------------------------------

# Populate year t of 2017-2018 transition year
# There are no 2018 data but this way we get all four years in the reproduction models
# Do this by creating the 2017-18 data as a stand-alone df then bind rows
LATR_dat_201718 <- LATR_full[LATR_full$year_t == 2016 & LATR_full$survival_t1 == 1, ]

# These are the 2017 survivors; make their year t demography last year's data
LATR_dat_201718$year_t <- 2017
LATR_dat_201718$year_t1 <- 2018
LATR_dat_201718$max.ht_t <- LATR_dat_201718$max.ht_t1
LATR_dat_201718$max.w_t <- LATR_dat_201718$max.w_t1
LATR_dat_201718$volume_t <- LATR_dat_201718$volume_t1
LATR_dat_201718$perp.w_t <- LATR_dat_201718$perp.w_t1
LATR_dat_201718$flowers_t <- LATR_dat_201718$flowers_t1
LATR_dat_201718$fruits_t <- LATR_dat_201718$fruits_t1
LATR_dat_201718$reproductive_fraction_t <- LATR_dat_201718$reproductive_fraction_t1
LATR_dat_201718$total.reproduction_t <- LATR_dat_201718$total.reproduction_t1

# Now set all the t1 data to NA
LATR_dat_201718$max.ht_t1 <- NA
LATR_dat_201718$max.w_t1 <- NA
LATR_dat_201718$volume_t1 <- NA
LATR_dat_201718$perp.w_t1 <- NA
LATR_dat_201718$flowers_t1 <- NA
LATR_dat_201718$fruits_t1 <- NA
LATR_dat_201718$reproductive_fraction_t1 <- NA
LATR_dat_201718$total.reproduction_t1 <- NA

# Bind rows and create log_vol as new variables (easier for GAMs)
LATR_flow_dat <- bind_rows(LATR_full, LATR_dat_201718) %>% 
  dplyr::select(unique.transect, volume_t, total.reproduction_t, weighted.dens) %>% drop_na()
LATR_flow_dat$log_volume_t <- log(LATR_flow_dat$volume_t)

# Create empty list to populate with model results
LATR_flower <- list()

# Three candidate models for the mean: size only, size + density, or size, density, and size:density
LATR_flower[[1]] <- gam(total.reproduction_t > 0 ~ s(log_volume_t) + s(unique.transect, bs = "re"),
                        data = LATR_flow_dat, gamma = gamma, family = "binomial")
LATR_flower[[2]] <- gam(total.reproduction_t > 0 ~ s(log_volume_t) + s(weighted.dens) + s(unique.transect, bs = "re"),
                        data = LATR_flow_dat, gamma = gamma, family = "binomial")
LATR_flower[[3]] <- gam(total.reproduction_t > 0 ~ s(log_volume_t) + s(weighted.dens) + ti(log_volume_t, weighted.dens) +
                          s(unique.transect, bs = "re"),
                        data = LATR_flow_dat, gamma = gamma, family = "binomial")

# Collect model AICs into a single table
flower_aic <- AICtab(LATR_flower, base = TRUE, sort = FALSE)

# Set top model as "best"
LATR_flower_best <- LATR_flower[[which.min(flower_aic$AIC)]]
LATR_flower_fitted_terms <- predict(LATR_flower_best, type = "terms") 
LATR_flow_dat$pred <- predict.gam(LATR_flower_best, newdata = LATR_flow_dat, exclude = "s(unique.transect)")





##### Fruit production model ------------------------------------------------------------------------------------------------------------------------

# Create new df with plants that have produced at least one reproductive structure
LATR_fruits_dat <- subset(LATR_flow_dat, total.reproduction_t > 0)

# Create empty list to populate with model results
LATR_fruits <- list()

# Three candidate models for the mean: size only, size + density, or size, density, and size:density
LATR_fruits[[1]] <- gam(total.reproduction_t ~ s(log_volume_t) + s(unique.transect, bs = "re"),
                        data = LATR_fruits_dat, gamma = gamma, family = "nb")
LATR_fruits[[2]] <- gam(total.reproduction_t ~ s(log_volume_t) + s(weighted.dens) + s(unique.transect, bs = "re"),
                        data = LATR_fruits_dat, gamma = gamma, family = "nb")
LATR_fruits[[3]] <- gam(total.reproduction_t ~ s(log_volume_t) + s(weighted.dens) + ti(log_volume_t, weighted.dens) +
                          s(unique.transect, bs = "re"),
                        data = LATR_fruits_dat, gamma = gamma, family = "nb")

# Collect model AICs into a single table
fruits_aic <- AICtab(LATR_fruits, base = TRUE, sort = FALSE)

# Set top model as "best"
LATR_fruits_best <- LATR_fruits[[which.min(fruits_aic$AIC)]]
LATR_fruits_fitted_terms <- predict(LATR_fruits_best, type = "terms") 
LATR_fruits_dat$pred <- predict.gam(LATR_fruits_best, newdata = LATR_fruits_dat, exclude = "s(unique.transect)")

# Plot effect of size on fruits
# plot(LATR_fruits_dat$log_volume_t, LATR_fruits_fitted_terms[, "s(log_volume_t)"]) 

# Plot effect of density on fruits 
# plot(LATR_fruits_dat$weighted.dens, LATR_fruits_fitted_terms[, "s(weighted.dens)"]) 





##### Survival model --------------------------------------------------------------------------------------------------------------------------------

# Combine transplants with large shrubs; keep only location info, survival, volume, and density
CData.Transplants %>% 
  dplyr::select("site", "transect", "actual.window", 
         "spring_survival_t1", "volume_t", "weighted.dens", "transplant") %>% 
  rename("survival_t1" = "spring_survival_t1") %>% 
  mutate(unique.transect = interaction(transect, site)) %>% 
  rbind(dplyr::select(LATR_full, "site", "transect", "actual.window", 
               "survival_t1", "volume_t", "weighted.dens", "transplant","unique.transect")) %>% 
  mutate(log_volume_t = log(volume_t)) %>% 
  drop_na() -> LATR_surv_dat

# Investigate size overlap between transplant experiment and observational census
# hist(log(LATR_surv_dat$volume_t[LATR_surv_dat$transplant == FALSE]), breaks = 25)
# hist(log(LATR_surv_dat$volume_t[LATR_surv_dat$transplant == TRUE]), breaks = 10,
#      add = TRUE, col = alpha("gray", 0.5))

# Plot survival against volume, grouped by transplant status
# plot(log(LATR_surv_dat$volume_t[LATR_surv_dat$transplant == FALSE]),
#      LATR_surv_dat$survival_t1[LATR_surv_dat$transplant == FALSE])
# points(log(LATR_surv_dat$volume_t[LATR_surv_dat$transplant == TRUE]),
#        LATR_surv_dat$survival_t1[LATR_surv_dat$transplant == TRUE] - 0.025, pch = 2)

# Create empty list to populate with model results
LATR_surv <- list()

# Three candidate models for the mean: size only, size + density, or size, density, and size:density
LATR_surv[[1]] <- gam(survival_t1 ~ s(log_volume_t,by = as.factor(transplant)) + s(unique.transect, bs = "re"),
                      data = LATR_surv_dat, gamma = gamma, family = "binomial")
LATR_surv[[2]] <- gam(survival_t1 ~ s(log_volume_t,by = as.factor(transplant)) + s(weighted.dens,by = as.factor(transplant)) +
                        s(unique.transect, bs = "re"),
                      data = LATR_surv_dat, gamma = gamma, family = "binomial")
LATR_surv[[3]] <- gam(survival_t1 ~ s(log_volume_t,by = as.factor(transplant)) + s(weighted.dens,by = as.factor(transplant)) +
                        ti(log_volume_t, weighted.dens, by = as.factor(transplant)) + s(unique.transect, bs = "re"),
                      data = LATR_surv_dat, gamma = gamma, family = "binomial")

# Collect model AICs into a single table
surv_aic <- AICtab(LATR_surv, base = TRUE, sort = FALSE)

# Set top model as "best"
LATR_surv_best <- LATR_surv[[which.min(surv_aic$AIC)]]
LATR_surv_fitted_terms <- predict(LATR_surv_best, type = "terms") 
LATR_surv_dat$pred <- predict.gam(LATR_surv_best, newdata = LATR_surv_dat, exclude = "s(unique.transect)")

# Plot effect of size on pr(survival)
# plot(LATR_surv_dat$log_volume_t, LATR_surv_fitted_terms[, "s(log_volume_t)"]) 

# Plot effect of density on pr(survival)
# plot(LATR_surv_dat$weighted.dens, LATR_surv_fitted_terms[, "s(weighted.dens)"]) 





##### Per-seed recruitment probability model --------------------------------------------------------------------------------------------------------

## number of seeds per fruit
seeds_per_fruit <- 5

# Create subset df of recruits
LATR_recruits <- LATR_full %>% 
  mutate(unique.transect = interaction(transect, site)) %>% 
  group_by(year_t1, unique.transect, actual.window) %>% 
  filter(seedling_t1 == 1)
  suppressMessages(summarise(LATR_recruits, recruits = n())) %>% 
  rename(window = actual.window) -> LATR_recruits

# Estimate total seeds produced in each window
# This is computed using the known plant sizes and the fitted flowering and fruiting models
# Note: we assume 5 seeds per fruit
LATR_transects <- Cdata.Transects.Windows %>% 
  mutate(unique.transect = interaction(transect, site),
         log_volume_t = log(volume))
LATR_transects$seeds = ceiling(invlogit(predict.gam(LATR_flower_best, newdata = LATR_transects))* 
                                 seeds_per_fruit*exp(predict.gam(LATR_fruits_best, newdata = LATR_transects)))
LATR_transects <- group_by(LATR_transects, unique.transect, window)
suppressMessages(summarise(LATR_transects, total_seeds = sum(seeds),
                                           weighted.dens = unique(weighted.dens))) -> LATR_transects

# Take three copies of this df, assigning each one to a different year and assigning recruits to zero (for now)
# Use only FPS for 2013-2014
LATR_recruitment <- bind_rows(LATR_transects %>% filter(unique.transect == "1.FPS" | unique.transect == "2.FPS" | unique.transect == "3.FPS") %>% 
                                mutate(year_t1 = 2014, recruits = 0),
                              LATR_transects %>% mutate(year_t1 = 2015, recruits = 0),
                              LATR_transects %>% mutate(year_t1 = 2016, recruits = 0),
                              LATR_transects %>% mutate(year_t1 = 2017, recruits = 0)) %>% 
  left_join(., LATR_recruits, by = c("year_t1", "unique.transect", "window")) %>% 
  mutate(recruits.y = replace_na(recruits.y, 0),
         recruits = pmax(recruits.x, recruits.y, na.rm = T)) %>% 
  drop_na()

# Create empty list to populate with model results
LATR_recruit <- list()

# Two candidate models for the mean: no effect, or weighted density only
LATR_recruit[[1]] <- gam(cbind(recruits,total_seeds - recruits) ~ s(unique.transect, bs = "re"),
                         data = LATR_recruitment, gamma = gamma, family = "binomial")
LATR_recruit[[2]] <- gam(cbind(recruits,total_seeds - recruits) ~ s(weighted.dens) + s(unique.transect, bs = "re"),
                         data = LATR_recruitment, gamma = gamma, family = "binomial")

# Collect model AICs into a single table
recruit_aic <- AICtab(LATR_recruit, base = TRUE, sort = FALSE)

# Set top model as "best"
LATR_recruit_best <- LATR_recruit[[which.min(recruit_aic$AIC)]]

# Plot effect of density on pr(seedling recruitment); negative density dependence
# LATR_recruit_fitted_terms <- predict(LATR_recruit[[2]], type = "terms") 
# plot(LATR_recruitment$weighted.dens, LATR_recruit_fitted_terms[, "s(weighted.dens)"])





##### Recruit sizes and integration limits (size bounds) --------------------------------------------------------------------------------------------

# Filter out seedlings from full data set; only do this once
if(boot.switch == FALSE){

  # Filter out seedlings and get their sizes
  LATR_recruit_size <- LATR_full %>% 
    filter(seedling_t1 == 1) %>% 
    mutate(log_volume = log(volume_t1)) %>% 
    arrange(weighted.dens)}

# Plot distribution of recruit sizes using hist(LATR_recruit_size$log_volume)
# hist(LATR_recruit_size$log_volume)

# Test for density dependence in recruit size
LATR_recruit_size_mod <- list()
LATR_recruit_size_mod[[1]] <- gam(list(log_volume ~ 1 + s(unique.transect, bs = "re"), ~ 1), 
                                  data = LATR_recruit_size, gamma = gamma, family = gaulss())
LATR_recruit_size_mod[[2]] <- gam(list(log_volume ~ s(weighted.dens) + s(unique.transect, bs = "re"), ~ 1), 
                                  data = LATR_recruit_size, gamma = gamma, family = gaulss())
LATR_recruit_size_mod[[3]] <- gam(list(log_volume ~ 1 + s(unique.transect, bs = "re"), ~ s(weighted.dens)), 
                                  data = LATR_recruit_size, gamma = gamma, family = gaulss())
LATR_recruit_size_mod[[4]] <- gam(list(log_volume ~ s(weighted.dens) + s(unique.transect, bs = "re"), ~ s(weighted.dens)), 
                                  data = LATR_recruit_size, gamma = gamma, family = gaulss())

# Collect model AICs into a single table
recruitsize_aic <- AICtab(LATR_recruit_size_mod, base = TRUE, sort = FALSE)

# Set top model as "best"
LATR_recruitsize_best <- LATR_recruit_size_mod[[which.min(recruitsize_aic$AIC)]]
LATR_recruitsize_fitted_terms <- predict(LATR_recruitsize_best, type = "terms") 
LATR_recruit_size$pred <- predict.gam(LATR_recruitsize_best, newdata = LATR_recruit_size, exclude = "s(unique.transect)")

# Annoying but necessary index wrangling; first line is where SD coefficients start
recruit_size_sd_index <- which(as.factor(names(coef(LATR_recruitsize_best))) == "(Intercept).1")
recruit_size_coef_length <- length(coef(LATR_recruitsize_best))

# Create maximum and minimum size bounds for the IPM
LATR_size_bounds <- data.frame(min_size = log(min(LATR_full$volume_t, LATR_full$volume_t1[LATR_full$transplant == FALSE], na.rm = TRUE)),
                               max_size = log(max(LATR_full$volume_t, LATR_full$volume_t1[LATR_full$transplant == FALSE], na.rm = TRUE)))





# Collect AIC table info ----------------------------------------------------------------------------------------------------------------------------

# Survival model AIC
surv_aic_out<-as.data.frame(surv_aic)
surv_aic_out$`Pr(Survival)` <- c("~ size + transplant + size:transplant + (1|transect)",
                                 "~ size + transplant + density + size:transplant + density:transplant + (1|transect)",
                                 "~ size + transplant + density + size:transplant + density:transplant + size:density + size:transplant:density + (1|transect)")

# Growth model AIC
grow_aic_out<-as.data.frame(grow_aic)
grow_aic_out$`mean(size)` <- c("~ size + (1|transect)", "~size + density + (1|transect)", "~ size + density + size:density + (1|transect)",
                               "~ size + (1|transect)", "~size + density + (1|transect)", "~ size + density + size:density + (1|transect)",
                               "~ size + (1|transect)", "~size + density + (1|transect)", "~ size + density + size:density + (1|transect)")
grow_aic_out$`sd(size)` <- c("~ 1", "~ 1", "~ 1",
                             "~ size", "~ size", "~ size",
                             "~ size + density", "~ size + density", "~ size + density")          

# Flowering model AIC
flower_aic_out <- as.data.frame(flower_aic)
flower_aic_out$`Pr(Flowering)` <- c("~ size + (1|transect)",
                                    "~ size + density + (1|transect)",
                                    "~ size + density + size:density + (1|transect)")

# Reproduction model AIC
fruits_aic_out <- as.data.frame(fruits_aic)
fruits_aic_out$`No. fruits` <- c("~ size + (1|transect)",
                                 "~ size + density + (1|transect)",
                                 "~ size + density + size:density + (1|transect)")

# Recruit model AIC
recruit_aic_out <- as.data.frame(recruit_aic)
recruit_aic_out$`Pr(Recruitment)` <- c("~ (1|transect)", "~ density + (1|transect)")

# Recruit size model AIC
recruitsize_aic_out<-as.data.frame(recruitsize_aic)
recruitsize_aic_out$`mean(size)` <- c("~ (1|transect)", "~ density + (1|transect)",
                                      "~ (1|transect)", "~ density + (1|transect)")
recruitsize_aic_out$`sd(size)` <- c("~1", "~ 1", "~ density", "~ density")

# Create list of AIC tables
aic_tables <- list(surv_aic_out = surv_aic_out,
                   grow_aic_out = grow_aic_out,
                   flower_aic_out = flower_aic_out,
                   fruits_aic_out = fruits_aic_out,
                   recruit_aic_out = recruit_aic_out,
                   recruitsize_aic_out = recruitsize_aic_out)
     
# Write out these quantities as an Rds file (do this only once)
# write_rds(x = aic_tables, file = "Manuscript/MS_Current/aic_tables.rds")
